Se busca bancos dado coordenadas y rango de distancia

+++ modelo/banco.go
@@ -6,8 +6,8 @@ type Banco struct {
 	Id          bson.ObjectId `json:"id" bson:"_id"`
 	ProvinciaId bson.ObjectId `json:"provincia" bson:"provinciaId"`
 	CiudadId    bson.ObjectId `json:"ciudad" bson:"ciudadId"`
-	Lat         float32       `json:"lat" bson:"lat"`
-	Lon         float32       `json:"lon" bson:"lon"`
+	Lat         float64       `json:"lat" bson:"lat"`
+	Lon         float64       `json:"lon" bson:"lon"`
 	Institución string        `json:"nombre" bson:"nombre"`
 	Dirección   string        `json:"direccion" bson:"direccion"`
 	Teléfono    string        `json:"telefono" bson:"telefono"`
@@ -17,7 +17,7 @@ type Banco struct {
 type Bancos []Banco
 
 func NewBanco(provinciaId, ciudadId bson.ObjectId,
-	lat, lon float32, institución, dirección, teléfono, horario string) *Banco {
+	lat, lon float64, institución, dirección, teléfono, horario string) *Banco {
 	return &Banco{
 		bson.NewObjectId(),
 		provinciaId,
